Ver. 0.9.4 ALPHA

+++ src/Controllers/BlublogCategoryController.php
@@ -4,82 +4,52 @@ namespace   Blublog\Blublog\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Cache;
 use App\Http\Controllers\Controller;
 use Blublog\Blublog\Models\Category;
 use Blublog\Blublog\Models\File;
 use Blublog\Blublog\Models\Post;
 use Blublog\Blublog\Models\Log;
+use Blublog\Blublog\Models\BlublogUser;
 use Session;
 
 class BlublogCategoryController extends Controller
 {
     public function index()
     {
+        BlublogUser::check_access('view', Category::class);
         $categories = Category::latest()->get();
-
-        // Getting total views of all post for all categories
-        // Useless for now
-        foreach($categories as $category){
-            $views = 0;
-            $numb = 1;
-            foreach($category->posts as $post ){
-
-                if( $numb <= $post->views->count()){
-                    $numb = $post->views->count();
-                    $category->mostviewsid = $post->id;
-                    $category->mostviewstitle = $post->title;
-                }
-                $views = $views + $post->views->count();
-            }
-            $category->views = $views;
-        }
-
         return view("blublog::panel.categories.index")->with('categories', $categories);
     }
     public function store(Request $request)
     {
+        BlublogUser::check_access('create', Category::class);
             $rules = [
                 'title' => 'required|max:255',
             ];
             $this->validate($request, $rules);
 
-            //Make slug from title
-            $slug = Post::makeslug($request->title);
-
         $category = new Category;
         $category->title = $request->title;
         $category->descr = $request->descr;
-        $category->slug = $slug;
+        $category->slug =  Post::makeslug($request->title);
         if($request->rgb){
             $category->colorcode = $request->rgb . ";";
         }else{
             $category->colorcode = "rgb(". rand(1,255) . "," . rand(1,255) . "," . rand(1,255) . ");";
         }
-
         if($request->file){
-                    $size = File::get_file_size($request->file);
-                    $numb = rand(0, 99);
-                    $numb2 = rand(99, 99);
-                    $numb = $numb . $numb2;
-                    $address = $numb . $request->file->getClientOriginalName();
-                    Storage::disk(config('blublog.files_disk', 'blublog'))->putFileAs('categories', $request->file, $address);
-
-                    $file = new File;
-                    $file->size = $size;
-                    $file->descr = __('files.image_for_category') . $request->title;
-                    $file->filename = 'categories/' . $address;
-                    $file->save();
-                    $category->img = $address;
+            $category->img = File::handle_img_upload_from_category($request);
         }
-
         $category->save();
-
+        Cache::forget('blublog.categories');
         Session::flash('success', __('blublog.contentcreate'));
         Log::add($request, "info", __('blublog.contentcreate') );
         return redirect()->route('blublog.categories.index');
     }
     public function edit($id)
     {
+        BlublogUser::check_access('update', Category::class);
         $category = Category::find($id);
         if(!$category){
             abort(404);
@@ -90,6 +60,7 @@ class BlublogCategoryController extends Controller
     }
     public function update(Request $request, $id)
     {
+        BlublogUser::check_access('update', Category::class);
         $rules = [
             'title' => 'required|max:255',
             'slug' => 'required|unique:blublog_posts|max:255',
@@ -100,26 +71,12 @@ class BlublogCategoryController extends Controller
         $category = Category::find($id);
 
         if($request->file){
-            $path = 'categories/' . $category->img;
-            $old_img = File::where([
-                ['filename', '=', $path],
-            ])->first();
-            $old_img->delete();
-            Storage::disk(config('blublog.files_disk', 'blublog'))->delete($path);
-            //GET THE SIZE OF FILE
-            $size = File::get_file_size($request->file);
-            $numb = rand(99, 9999);
-            $address =  $numb . $request->file->getClientOriginalName();
-            Storage::disk(config('blublog.files_disk', 'blublog'))->putFileAs('categories', $request->file, $address);
-
-            $file = new File;
-            $file->size = $size;
-            $file->descr = __('files.image_for_category') . $request->title;
-            $file->filename = 'categories/' . $address;
-            $file->save();
-
-
-
+            $old_img = File::get_category_img_file($category->img);
+            if($old_img){
+                Storage::disk(config('blublog.files_disk', 'blublog'))->delete($path);
+                $old_img->delete();
+            }
+            $address = File::handle_img_upload_from_category($request);
         } else {
             $address = $category->img;
         }
@@ -134,23 +91,19 @@ class BlublogCategoryController extends Controller
         $category->slug = $request->slug;
         $category->img = $address;
         $category->save();
-
+        Cache::forget('blublog.categories');
         Log::add($request, "info", __('blublog.contentupdate') );
-        Session::flash('success', __('general.contentupdate'));
+        Session::flash('success', __('blublog.contentupdate'));
         return redirect()->back();
     }
     public function destroy($id)
     {
-
-
-
+        BlublogUser::check_access('delete', Category::class);
         $Category = Category::find($id);
         if(!$Category){
             Session::flash('error', __('general.content_does_not_found'));
             return redirect()->route('categories.index');
         }
-
-
 
         $path = 'categories/' . $Category->img;
         $file = File::where([
@@ -158,7 +111,9 @@ class BlublogCategoryController extends Controller
         ])->first();
         if($file){
             $file->delete();
-            Storage::disk(config('blublog.files_disk', 'blublog'))->delete($path);
+            if(!Storage::disk(config('blublog.files_disk', 'blublog'))->delete($path)){
+                Log::add($Category, "error", __('blublog.error_removing') );
+            }
         }
         $Category->posts()->detach();
         Log::add($Category, "info", __('blublog.contentdelete') );
